+++ wgpu-hal/src/metal/mod.rs
@@ -33,6 +33,7 @@ use std::{
 };
 
 use arrayvec::ArrayVec;
+use bitflags::bitflags;
 use metal::foreign_types::ForeignTypeRef as _;
 use parking_lot::Mutex;
 
@@ -145,6 +146,24 @@ impl crate::Instance<Api> for Instance {
     }
 }
 
+bitflags!(
+    /// Similar to `MTLCounterSamplingPoint`, but a bit higher abstracted for our purposes.
+    #[derive(Debug, Copy, Clone)]
+    pub struct TimestampQuerySupport: u32 {
+        /// On creating Metal encoders.
+        const STAGE_BOUNDARIES = 1 << 1;
+        /// Within existing draw encoders.
+        const ON_RENDER_ENCODER = Self::STAGE_BOUNDARIES.bits() | (1 << 2);
+        /// Within existing dispatch encoders.
+        const ON_COMPUTE_ENCODER = Self::STAGE_BOUNDARIES.bits() | (1 << 3);
+        /// Within existing blit encoders.
+        const ON_BLIT_ENCODER = Self::STAGE_BOUNDARIES.bits() | (1 << 4);
+
+        /// Within any wgpu render/compute pass.
+        const INSIDE_WGPU_PASSES = Self::ON_RENDER_ENCODER.bits() | Self::ON_COMPUTE_ENCODER.bits();
+    }
+);
+
 #[allow(dead_code)]
 #[derive(Clone, Debug)]
 struct PrivateCapabilities {
@@ -241,8 +260,7 @@ struct PrivateCapabilities {
     supports_preserve_invariance: bool,
     supports_shader_primitive_index: bool,
     has_unified_memory: Option<bool>,
-    support_timestamp_query: bool,
-    support_timestamp_query_in_passes: bool,
+    timestamp_query_support: TimestampQuerySupport,
 }
 
 #[derive(Clone, Debug)]
@@ -706,7 +724,7 @@ pub struct ComputePipeline {
 unsafe impl Send for ComputePipeline {}
 unsafe impl Sync for ComputePipeline {}
 
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct QuerySet {
     raw_buffer: metal::Buffer,
     //Metal has a custom buffer for counters.
@@ -789,6 +807,9 @@ struct CommandState {
 
     work_group_memory_sizes: Vec<u32>,
     push_constants: Vec<u32>,
+
+    /// Timer query that should be executed when the next pass starts.
+    pending_timer_queries: Vec<(QuerySet, u32)>,
 }
 
 pub struct CommandEncoder {